+++ nodes/Fiveday/TaskLinkDescription.ts
@@ -1,35 +1,35 @@
 import type { INodeProperties } from 'n8n-workflow';
 
-export const taskTagOperations: INodeProperties = {
+export const taskLinkOperations: INodeProperties = {
 	displayName: 'Operation',
 	name: 'operation',
 	type: 'options',
 	noDataExpression: true,
 	displayOptions: {
 		show: {
-			resource: ['taskTag'],
+			resource: ['taskLink'],
 		},
 	},
 	options: [
 		{
-			name: 'Add Tag',
-			value: 'addTag',
-			description: 'Add a task tag',
-			action: 'Add a task tag',
+			name: 'Add Link',
+			value: 'addLink',
+			description: 'Add a link to a task',
+			action: 'Add a link to a task',
 		},
 		{
-			name: 'Remove Tag',
-			value: 'removeTag',
-			description: 'Remove a task tag',
-			action: 'Remove a task tag',
+			name: 'Remove Link',
+			value: 'removeLink',
+			description: 'Remove a link from a task',
+			action: 'Remove a link from a task',
 		},
 	],
-	default: 'addTag',
+	default: 'addLink',
 };
 
-export const taskTagFields: INodeProperties[] = [
+export const taskLinkFields: INodeProperties[] = [
 	// ----------------------------------
-	//         Task Tag: Add Tag
+	//         Task Link: Add Link
 	// ----------------------------------
 	{
 		displayName: 'Task ID',
@@ -40,11 +40,11 @@ export const taskTagFields: INodeProperties[] = [
 		placeholder: 'e.g. PRO-1',
 		displayOptions: {
 			show: {
-				resource: ['taskTag'],
-				operation: ['addTag'],
+				resource: ['taskLink'],
+				operation: ['addLink'],
 			},
 		},
-		description: 'The ID of the task to add a tag to',
+		description: 'The ID of the task to add a link to',
 	},
 	{
 		displayName: 'Project Name or ID',
@@ -58,32 +58,32 @@ export const taskTagFields: INodeProperties[] = [
 		required: true,
 		displayOptions: {
 			show: {
-				resource: ['taskTag'],
-				operation: ['addTag'],
+				resource: ['taskLink'],
+				operation: ['addLink'],
 			},
 		},
 		description: 'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
 	},
 	{
-		displayName: 'Tag Names or IDs',
-		name: 'tags',
+		displayName: 'Link Task Names or IDs',
+		name: 'linkedTaskIds',
 		type: 'multiOptions',
 		typeOptions: {
-			loadOptionsMethod: 'getTags',
-			loadOptionsDependsOn: ['projectId'],
+			loadOptionsMethod: 'getTasksForLink',
+			loadOptionsDependsOn: ['projectId', 'taskId'],
 		},
 		default: [],
 		required: true,
 		displayOptions: {
 			show: {
-				resource: ['taskTag'],
-				operation: ['addTag'],
+				resource: ['taskLink'],
+				operation: ['addLink'],
 			},
 		},
 		description: 'Choose from the list, or specify IDs using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
 	},
 	// ----------------------------------
-	//         Task Tag: Remove Tag
+	//        Task Link: Remove Link
 	// ----------------------------------
 	{
 		displayName: 'Task ID',
@@ -94,11 +94,11 @@ export const taskTagFields: INodeProperties[] = [
 		placeholder: 'e.g. PRO-1',
 		displayOptions: {
 			show: {
-				resource: ['taskTag'],
-				operation: ['removeTag'],
+				resource: ['taskLink'],
+				operation: ['removeLink'],
 			},
 		},
-		description: 'The ID of the task to remove a tag from',
+		description: 'The ID of the task to remove a link from',
 	},
 	{
 		displayName: 'Project Name or ID',
@@ -112,26 +112,26 @@ export const taskTagFields: INodeProperties[] = [
 		required: true,
 		displayOptions: {
 			show: {
-				resource: ['taskTag'],
-				operation: ['removeTag'],
+				resource: ['taskLink'],
+				operation: ['removeLink'],
 			},
 		},
 		description: 'Choose from the list, or specify an ID using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',
 	},
 	{
-		displayName: 'Tag Names or IDs',
-		name: 'tags',
+		displayName: 'Link Task Names or IDs',
+		name: 'linkedTaskIds',
 		type: 'multiOptions',
 		typeOptions: {
-			loadOptionsMethod: 'getTaskTags',
-			loadOptionsDependsOn: ['taskId'],
+			loadOptionsMethod: 'getTasksForLink',
+			loadOptionsDependsOn: ['projectId', 'taskId'],
 		},
 		default: [],
 		required: true,
 		displayOptions: {
 			show: {
-				resource: ['taskTag'],
-				operation: ['removeTag'],
+				resource: ['taskLink'],
+				operation: ['removeLink'],
 			},
 		},
 		description: 'Choose from the list, or specify IDs using an <a href="https://docs.n8n.io/code/expressions/">expression</a>',